backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
Pillow
psycopg2-binary
python-decouple
whitenoise
